✨ 면접 관리에 필요한 Entity 생성 로직

+++ src/main/java/KUSITMS/WITHUS/domain/user/repository/UserJpaRepository.java
@@ -3,7 +3,7 @@ package KUSITMS.WITHUS.domain.user.repository;
 import KUSITMS.WITHUS.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserJpaRepository extends JpaRepository<User, Integer> {
+public interface UserJpaRepository extends JpaRepository<User, Long> {
     Boolean existsByEmail(String email);
     User findByEmail(String email);
 }
